feat: properly parse the reclaim script lock time (#713)
* Properly parse locktimes.
* Update emily and signer after the locktime change
* Update the docstrings

+++ emily/handler/src/api/models/deposit/mod.rs
@@ -79,7 +79,7 @@ pub struct DepositParameters {
     /// the transaction.
     pub max_fee: u64,
     /// Bitcoin block height at which the reclaim script becomes executable.
-    pub lock_time: u64,
+    pub lock_time: u32,
 }
 
 /// Reduced version of the Deposit data.
